fix: phpstan / testing issues

+++ app/Livewire/Settings/DeleteUserForm.php
@@ -21,7 +21,9 @@ final class DeleteUserForm extends Component
             'password' => ['required', 'string', 'current_password'],
         ]);
 
-        tap(Auth::user(), $logout(...))->delete();
+        if (Auth::user()) {
+            tap(Auth::user(), $logout(...))->delete();
+        }
 
         $this->redirect('/', navigate: true);
     }
